Wacks/fixes all warnings on Solaris, MacOS and FreeBSD. Tested against OpenSSL only
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@89208 13f79535-47bb-0310-9956-ffa450edef68

+++ src/support/logresolve.c
@@ -314,7 +314,7 @@ int main (int argc, char *argv[])
 	if (line[0] == '\0')
 	    continue;
 	entries++;
-	if (!isdigit(line[0])) {	/* short cut */
+	if (!isdigit((int)line[0])) {	/* short cut */
 	    puts(line);
 	    withname++;
 	    continue;
